Fixed GRASSConfig constructor to have both newLocation and overwrite flags. If newLocation is True, will raise an error if the location directory already exists, unless overwrite is True, in which case the location directory will be deleted

+++ ecohydrolib/grasslib.py
@@ -43,14 +43,15 @@ DEFAULT_LOCATION = 'default'
 DEFAULT_MAPSET = 'PERMANENT'
 
 class GRASSConfig(object):
-    def __init__(self, context, dbase, location=None, mapset=None, overwrite=False):
+    def __init__(self, context, dbase, location=None, mapset=None, newLocation=False, overwrite=False):
         """ @brief Constructor for GRASSConfig
         
             @param context ecohydrolib.Context with config containing a GRASS section with GISBASE entry
             @param dbase String representing a GRASS GIS data directory
             @param location String representing a GRASS project location within the data directory; defaults to DEFAULT_LOCATION.
             @param mapset String representing a GRASS mapset within the location; defaults to DEFAULT_MAPSET.
-            @param overwrite Boolean, if False, constructor will raise an error if the location already exists
+            @param newLocation Boolean, True if this is a new location, False if the location is expected to already exist
+            @param overwrite Boolean, if False, constructor will raise an error if newLocation is True and the location already exists
             
             @raise IOError(errno.ENOTDIR) if dbase is not a writable directory
             @raise IOError(errno.EACCESS) if dbase is not writable
@@ -76,12 +77,13 @@ class GRASSConfig(object):
     
         locationPath = os.path.join(self.dbase, self.location)
         if os.path.exists( locationPath ):
-            if overwrite:
-                # Delete existing location
-                shutil.rmtree(locationPath)
-            else:
-                raise IOError(errno.EEXIST, "Location '%s' already exists in %s" % \
-                              (location, self.dbase))
+            if newLocation:
+                if overwrite:
+                    # Delete existing location
+                    shutil.rmtree(locationPath)
+                else:
+                    raise IOError(errno.EEXIST, "Location '%s' already exists in %s" % \
+                                  (location, self.dbase))
         self.mapset = mapset
 
 class GRASSLib(object):
